refactor: remove stateUpdate and boardcast
state update should be handled by uSESW
