+++ src/utils/global-state.ts
@@ -1,13 +1,11 @@
 import {
   Cache,
   ScopedMutator,
-  RevalidateCallback,
-  StateUpdateCallback
+  RevalidateCallback
 } from '../types'
 
 export type GlobalState = [
   Record<string, RevalidateCallback[]>, // EVENT_REVALIDATORS
-  Record<string, StateUpdateCallback[]>, // STATE_UPDATERS
   Record<string, [number, number]>, // MUTATION: [ts, end_ts]
   Record<string, [any, number]>, // FETCH: [data, ts]
   ScopedMutator, // Mutator
